+++ packages/react-tinacms-editor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.52.8
+
+### Patch Changes
+
+- Updated dependencies [f6876d30f]
+- Updated dependencies [92268fc85]
+  - @tinacms/toolkit@0.56.1
+
 ## 0.52.7
 
 ### Patch Changes